neovim: Lazy migration structure and plugins version updates
