+++ vim/nvim-lua/init.lua
@@ -1,3 +1,6 @@
+----------
+-- Lazy --
+----------
 -- Bootstrap lazy.nvim
 local lazypath = vim.fn.stdpath("data") .. "/lazy/lazy.nvim"
 if not (vim.uv or vim.loop).fs_stat(lazypath) then
@@ -25,5 +28,40 @@ vim.o.termguicolors = true
 
 require('lazy').setup('plugins')
 
--- require('settings')
--- require('mappings')
+--------------
+-- Settings --
+--------------
+
+require('settings')
+
+
+--- Colorscheme
+
+local function colorize(color)
+  color = color or "base16-gruvbox-dark-pale"
+  vim.cmd.colorscheme(color)
+  -- vim.cmd([[colorscheme base16-everforest]])
+  -- vim.cmd([[colorscheme base16-material-ocean]])
+  -- vim.cmd([[colorscheme base16-material-palenight]])
+  -- vim.cmd([[colorscheme base16-tomorrow-night]])
+  -- vim.cmd([[colorscheme base16-eighties]])
+  -- vim.cmd([[colorscheme base16-base16-gruvbox-dark-pale]])
+
+  -- Set transparent background
+  vim.cmd('highlight Normal guibg=none')
+  vim.cmd('highlight NonText guibg=none')
+  vim.cmd('highlight Normal ctermbg=none')
+  vim.cmd('highlight NonText ctermbg=none')
+  vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
+  vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
+end
+
+colorize('catppuccin')
+
+
+
+--------------
+-- Mappings --
+--------------
+
+require('mappings')
