--- vim/nvim-lua/lua/configs/gitsigns.lua
@@ -1,2 +1,0 @@
-
-require('gitsigns').setup()
