--- vim/nvim-lua/lua/configs/nvim-colorizer.lua
@@ -1,6 +1,0 @@
-require('colorizer').setup({
-  '*',    -- Highlight all files, but customize some others.
-  '!vim', -- Exclude vim from highlighting.
-  css = { css = true },
-  html = { mode = 'background' },
-})
