--- vim/nvim-lua/lua/configs/test.lua
@@ -1,15 +1,0 @@
-require('neotest').setup({
-  adapters = {
-    require('neotest-vim-test')({
-      ignore_file_types = {
-        'vim',
-        'lua'
-      },
-      allow_file_types = {
-        'ts'
-        --   'haskell',
-        --   'elixir'
-      }
-    })
-  }
-})
