+++ vim/nvim-lua/lua/plugins/lualine.lua
@@ -1,78 +1,78 @@
-local colors = {
-  red = '#ca1243',
-  black = '#383a42',
-  white = '#f3f3f3',
-  light_green = '#83a598',
-  orange = '#fe8019',
-  green = '#8ec07c',
-  section_color = '#5f6a8f',
-  dark_grey = '#323c41',
-}
-
-local theme = {
-  normal = {
-    a = { fg = colors.white, bg = colors.black },
-    b = { fg = colors.white, bg = colors.section_color },
-    c = { fg = colors.black, bg = colors.dark_grey },
-    z = { fg = colors.white, bg = colors.black },
-  },
-  insert = { a = { fg = colors.black, bg = colors.light_green } },
-  visual = { a = { fg = colors.black, bg = colors.orange } },
-  replace = { a = { fg = colors.black, bg = colors.green } },
-}
-
-local empty = require('lualine.component'):extend()
-function empty:draw(default_highlight)
-  self.status = ''
-  self.applied_separator = ''
-  self:apply_highlights(default_highlight)
-  self:apply_section_separators()
-  return self.status
-end
-
--- Put proper separators and gaps between components in sections
-local function process_sections(sections)
-  for name, section in pairs(sections) do
-    local left = name:sub(9, 10) < 'x'
-    for pos = 1, name ~= 'lualine_z' and #section or #section - 1 do
-      table.insert(section, pos * 2, { empty, color = { fg = colors.white, bg = colors.dark_grey } })
-    end
-    for id, comp in ipairs(section) do
-      if type(comp) ~= 'table' then
-        comp = { comp }
-        section[id] = comp
-      end
-      comp.separator = left and { right = '' } or { left = '' }
-    end
-  end
-  return sections
-end
-
-local function search_result()
-  if vim.v.hlsearch == 0 then
-    return ''
-  end
-  local last_search = vim.fn.getreg('/')
-  if not last_search or last_search == '' then
-    return ''
-  end
-  local searchcount = vim.fn.searchcount { maxcount = 9999 }
-  return last_search .. '(' .. searchcount.current .. '/' .. searchcount.total .. ')'
-end
-
-local function modified()
-  if vim.bo.modified then
-    return '+'
-  elseif vim.bo.modifiable == false or vim.bo.readonly == true then
-    return '-'
-  end
-  return ''
-end
-
 return {
   'nvim-lualine/lualine.nvim',
   dependencies = { 'kyazdani42/nvim-web-devicons', lazy = true },
   config = function()
+    local colors = {
+      red = '#ca1243',
+      black = '#383a42',
+      white = '#f3f3f3',
+      light_green = '#83a598',
+      orange = '#fe8019',
+      green = '#8ec07c',
+      section_color = '#5f6a8f',
+      dark_grey = '#323c41',
+    }
+
+    local theme = {
+      normal = {
+        a = { fg = colors.white, bg = colors.black },
+        b = { fg = colors.white, bg = colors.section_color },
+        c = { fg = colors.black, bg = colors.dark_grey },
+        z = { fg = colors.white, bg = colors.black },
+      },
+      insert = { a = { fg = colors.black, bg = colors.light_green } },
+      visual = { a = { fg = colors.black, bg = colors.orange } },
+      replace = { a = { fg = colors.black, bg = colors.green } },
+    }
+
+    local empty = require('lualine.component'):extend()
+    function empty:draw(default_highlight)
+      self.status = ''
+      self.applied_separator = ''
+      self:apply_highlights(default_highlight)
+      self:apply_section_separators()
+      return self.status
+    end
+
+    -- Put proper separators and gaps between components in sections
+    local function process_sections(sections)
+      for name, section in pairs(sections) do
+        local left = name:sub(9, 10) < 'x'
+        for pos = 1, name ~= 'lualine_z' and #section or #section - 1 do
+          table.insert(section, pos * 2, { empty, color = { fg = colors.white, bg = colors.dark_grey } })
+        end
+        for id, comp in ipairs(section) do
+          if type(comp) ~= 'table' then
+            comp = { comp }
+            section[id] = comp
+          end
+          comp.separator = left and { right = '' } or { left = '' }
+        end
+      end
+      return sections
+    end
+
+    local function search_result()
+      if vim.v.hlsearch == 0 then
+        return ''
+      end
+      local last_search = vim.fn.getreg('/')
+      if not last_search or last_search == '' then
+        return ''
+      end
+      local searchcount = vim.fn.searchcount { maxcount = 9999 }
+      return last_search .. '(' .. searchcount.current .. '/' .. searchcount.total .. ')'
+    end
+
+    local function modified()
+      if vim.bo.modified then
+        return '+'
+      elseif vim.bo.modifiable == false or vim.bo.readonly == true then
+        return '-'
+      end
+      return ''
+    end
+
     require('lualine').setup(
       {
         options = {
@@ -150,5 +150,5 @@ return {
         },
       }
     )
-  end
+  end,
 }
