+++ vim/nvim-lua/lua/plugins/nvim-colorizer.lua
@@ -1,0 +1,9 @@
+return {
+  'norcalli/nvim-colorizer.lua',
+  opts = {
+    '*',    -- Highlight all files, but customize some others.
+    '!vim', -- Exclude vim from highlighting.
+    css = { css = true },
+    html = { mode = 'background' },
+  }
+}
