+++ vim/nvim-lua/lua/settings.lua
@@ -1,74 +1,7 @@
-local g = vim.g
 local o = vim.o
 
 -- cmd('syntax on')
 -- vim.api.nvim_command('filetype plugin indent on')
-
-require('catppuccin').setup({
-  flavour = 'macchiato', -- latte, frappe, macchiato, mocha
-  background = {         -- :h background
-    light = 'latte',
-    dark = 'mocha',
-  },
-  transparent_background = false, -- disables setting the background color.
-  show_end_of_buffer = false,     -- shows the '~' characters after the end of buffers
-  term_colors = false,            -- sets terminal colors (e.g. `g:terminal_color_0`)
-  dim_inactive = {
-    enabled = true,               -- dims the background color of inactive window
-    shade = 'dark',
-    percentage = 0.15,            -- percentage of the shade to apply to the inactive window
-  },
-  no_italic = false,              -- Force no italic
-  no_bold = false,                -- Force no bold
-  no_underline = false,           -- Force no underline
-  styles = {                      -- Handles the styles of general hi groups (see `:h highlight-args`):
-    comments = { 'italic' },      -- Change the style of comments
-    conditionals = { 'italic' },
-    loops = {},
-    functions = {},
-    keywords = {},
-    strings = {},
-    variables = {},
-    numbers = {},
-    booleans = {},
-    properties = {},
-    types = {},
-    operators = {},
-  },
-  color_overrides = {},
-  custom_highlights = {},
-  integrations = {
-    cmp = true,
-    gitsigns = true,
-    nvimtree = true,
-    treesitter = true,
-    notify = false,
-    mini = false,
-    -- For more plugins integrations please scroll down (https://github.com/catppuccin/nvim#integrations)
-  },
-})
-
--- Colorscheme
-local function colorize(color)
-  color = color or "base16-gruvbox-dark-pale"
-  vim.cmd.colorscheme(color)
-  -- vim.cmd([[colorscheme base16-everforest]])
-  -- vim.cmd([[colorscheme base16-material-ocean]])
-  -- vim.cmd([[colorscheme base16-material-palenight]])
-  -- vim.cmd([[colorscheme base16-tomorrow-night]])
-  -- vim.cmd([[colorscheme base16-eighties]])
-  -- vim.cmd([[colorscheme base16-base16-gruvbox-dark-pale]])
-
-  -- Set transparent background
-  vim.cmd('highlight Normal guibg=none')
-  vim.cmd('highlight NonText guibg=none')
-  vim.cmd('highlight Normal ctermbg=none')
-  vim.cmd('highlight NonText ctermbg=none')
-  vim.api.nvim_set_hl(0, "Normal", { bg = "none" })
-  vim.api.nvim_set_hl(0, "NormalFloat", { bg = "none" })
-end
-
-colorize('catppuccin')
 
 o.background = 'dark'
